Use pytest for tests and split tests.py into multiple files
