+++ requirements-dev.txt
@@ -1,9 +1,7 @@
 -r requirements.txt
 
 # testing (e.g. `python setup.py -q test -v`)
-flake8
-mock
-unittest2
+pytest
 tox
 
 # docs generation (e.g. `cd docs && make html`)
