+++ setup.cfg
@@ -7,4 +7,4 @@ show_skipped = no
 exclude = docs/*
 
 [pytest]
-python_files = tests.py
+testpaths = tests/
