+++ tox.ini
@@ -18,8 +18,7 @@ basepython =
     py33: python3.3
     py34: python3.4
 deps =
-    mock
-    unittest2
+    pytest
     django14: Django>=1.4,<1.5
     django15: Django>=1.5,<1.6
     django16: Django>=1.6,<1.7
@@ -27,9 +26,10 @@ deps =
     django18: Django>=1.8,<1.9
     py26: importlib
     py26: argparse
+    py{26,27}: mock
 commands =
     {envpython} manage.py behave --tags=~@failing --format=progress
-    {envpython} tests.py
+    py.test
 
 [testenv:docs]
 basepython = python
